only ptask is missing and we can start with some heavy testing :-), but there is still a lot of bug I'd bet,... and I also has to check for both reviews from Kuba

+++ sssd/src/providers/ipa/ipa_sudo_export.c
@@ -505,9 +505,9 @@ fail:
 errno_t ipa_sudo_export_sudoers(TALLOC_CTX *mem, 
                                 struct sysdb_ctx *sysdb,
                                 struct sysdb_attrs **ipa_rules, 
-                                int rules_count, 
+                                size_t rules_count, 
                                 struct sysdb_attrs ***exported_rules,
-                                int *sudoers_count,
+                                size_t *sudoers_count,
                                 struct ipa_sudoer_cmds ***index,
                                 struct tevent_req *req)
 {
@@ -614,6 +614,7 @@ errno_t ipa_sudo_export_sudoers(TALLOC_CTX *mem,
 
               
             } /* EXPORT all values of the attribute */
+            // =================================================================
 
 #ifdef old
             /* EXPORT all values of the attribute */
